♻️  Split connection::Call into a separate file
