+++ zlink/src/connection/mod.rs
@@ -6,6 +6,8 @@ pub mod socket;
 mod write_connection;
 use core::fmt::Debug;
 pub use write_connection::WriteConnection;
+mod call;
+pub use call::Call;
 
 use serde::{Deserialize, Serialize};
 pub use socket::Socket;
@@ -154,41 +156,6 @@ impl<Params> Reply<Params> {
     /// If there are more replies to come.
     pub fn continues(&self) -> Option<bool> {
         self.continues
-    }
-}
-
-/// A method call.
-#[derive(Debug, Serialize, Deserialize)]
-pub struct Call<M> {
-    #[serde(flatten)]
-    method: M,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    oneway: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    more: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    upgrade: Option<bool>,
-}
-
-impl<M> Call<M> {
-    /// The method call name and parameters.
-    pub fn method(&self) -> &M {
-        &self.method
-    }
-
-    /// If the method call doesn't want a reply.
-    pub fn oneway(&self) -> Option<bool> {
-        self.oneway
-    }
-
-    /// If the method call is requesting more replies.
-    pub fn more(&self) -> Option<bool> {
-        self.more
-    }
-
-    /// If the method call is requesting an upgrade to a different protocol.
-    pub fn upgrade(&self) -> Option<bool> {
-        self.upgrade
     }
 }
